put itxs in instructions folder, add claimPending itx
